Addition

+++ src/Addition/UndirectedGraphNode.java
@@ -3,11 +3,11 @@ package Addition;
 import java.util.*;
 
 public class UndirectedGraphNode {
-	int value;
+	double value;
 	boolean known;
 	Map<UndirectedGraphNode, Integer> positiveNeighbors;
 	Map<UndirectedGraphNode, Integer> negativeNeighbors;
-	UndirectedGraphNode(String l, int x) { 
+	UndirectedGraphNode(String l, double x) { 
 		value = x; 
 		known = false;
 		positiveNeighbors = new HashMap<UndirectedGraphNode, Integer>();
